+++ scripts/details-popup/details-popup.js
@@ -41,7 +41,7 @@
 
 				if (request.popupOpacity) {
 					console.log(request.popupOpacity);
-					document.documentElement.style.setProperty('--detailsPopup-opacity', request.popupOpacity/10);
+					document.documentElement.style.setProperty('--detailsPopup-opacity', request.popupOpacity > 1 ? request.popupOpacity/10 : request.popupOpacity);
 				}
 			});