+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Popup
 - Slider on settings to change width of extension popup
 - Precise search switch button now works on kana writing and kanji (before, it was always precise)
+- Sliders on settings now work with mouse wheel too (shift + wheel increases scale)
 
 ## Bug Fix
 - Clicking on a search result while on kana writing mode now searches for that result (before, it only worked while on romaji mode)